@@ -34,18 +34,18 @@
   "dependencies": {
     "@nextcloud/auth": "^2.2.1",
     "@nextcloud/axios": "^2.4.0",
-    "@nextcloud/dialogs": "^5.1.2",
-    "@nextcloud/initial-state": "^2.1.0",
-    "@nextcloud/l10n": "^2.2.0",
+    "@nextcloud/dialogs": "^7.2.0",
+    "@nextcloud/initial-state": "^3.0.0",
+    "@nextcloud/l10n": "^3.4.1",
     "@nextcloud/moment": "^1.3.1",
-    "@nextcloud/paths": "^2.1.0",
+    "@nextcloud/paths": "^3.0.0",
     "@nextcloud/router": "^3.0.0",
-    "@nextcloud/vue": "^8.8.1",
-    "vue": "^2.7.16",
+    "@nextcloud/vue": "^9.3.1",
+    "vue": "^3.5.26",
     "vue-click-outside": "^1.1.0",
     "vue-clipboard2": "^0.3.3",
     "vue-material-design-icons": "^5.3.0",
-    "@nextcloud/password-confirmation": "^5.1.1"
+    "@nextcloud/password-confirmation": "^6.0.2"
   },
   "engines": {
     "node": "^20.0.0",
@@ -55,9 +55,9 @@
     "@nextcloud/babel-config": "^1.0.0",
     "@nextcloud/browserslist-config": "^3.0.0",
     "@nextcloud/eslint-config": "^8.3.0",
-    "@nextcloud/stylelint-config": "^2.1.2",
-    "@nextcloud/webpack-vue-config": "^6.0.0",
-    "eslint-webpack-plugin": "^4.0.0",
-    "stylelint-webpack-plugin": "^4.0.0"
+    "@nextcloud/stylelint-config": "^3.1.1",
+    "@nextcloud/webpack-vue-config": "^7.0.2",
+    "eslint-webpack-plugin": "^5.0.2",
+    "stylelint-webpack-plugin": "^5.0.1"
   }
 }
