@@ -1,19 +1,12 @@
-/* jshint esversion: 6 */
-
 /**
  * SPDX-FileCopyrightText: 2020 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import Vue from 'vue'
-import './bootstrap.js'
+import { createApp } from 'vue'
 import AdminSettings from './components/AdminSettings.vue'
 
-// eslint-disable-next-line
-'use strict'
+const app = createApp(AdminSettings)
 
-// eslint-disable-next-line
-new Vue({
-	el: '#nuiteq_prefs',
-	render: h => h(AdminSettings),
-})
+app.mixin({ methods: { t, n } })
+app.mount('#nuiteq_prefs')
