@@ -28,8 +28,8 @@
 
 <script>
 import PlusIcon from 'vue-material-design-icons/Plus.vue'
-import NcAppNavigationNew from '@nextcloud/vue/dist/Components/NcAppNavigationNew.js'
-import NcAppNavigation from '@nextcloud/vue/dist/Components/NcAppNavigation.js'
+import NcAppNavigationNew from '@nextcloud/vue/components/NcAppNavigationNew'
+import NcAppNavigation from '@nextcloud/vue/components/NcAppNavigation'
 
 import BoardNavigationItem from './BoardNavigationItem.vue'
 
